Move tcp.endpoint.waiterQueue initialization for passively connected endpoints.
According to the field's contract, it must be initialized at endpoint creation
and not change throughout the lifetime of the endpoint. This change makes that
actually true. Previously, passively connected endpoints would initialize their
waiterQueue field after endpoint creation.

This was a problem when calling stack.Close when endpoints were still being
accepted.

PiperOrigin-RevId: 292013340